requirements.txt
datadog==0.31.0
paho-mqtt==1.5.0
requests==2.22.0
